requirements.txt
dnspython==2.3.0
sublistr3r==1.0
termcolor==2.3.0
Flask==2.1.0